+++ eventstore/test_views.py
@@ -27,6 +27,7 @@ from eventstore.models import (
     ChannelSwitch,
     CHWRegistration,
     Covid19Triage,
+    DeliveryFailure,
     EddSwitch,
     Event,
     HealthCheckUserProfile,
@@ -637,6 +638,88 @@ class PrebirthRegistrationViewSetTests(APITestCase, BaseEventTestCase):
         self.assertEqual(registration.facility_code, "123456")
         self.assertEqual(registration.source, "WhatsApp")
         self.assertEqual(registration.created_by, user.username)
+
+        self.assertFalse(DeliveryFailure.objects.all().exists())
+
+    @responses.activate
+    def test_reset_delivery_failures(self):
+        """
+        Should create a new PrebirthRegistration object in the database and reset
+        the delivery failure record if present
+        """
+        contact_uuid = "9e12d04c-af25-40b6-aa4f-57c72e8e3f91"
+        wa_id = "27820001001"
+
+        user = get_user_model().objects.create_user("test")
+        user.user_permissions.add(
+            Permission.objects.get(codename="add_prebirthregistration")
+        )
+        self.client.force_authenticate(user)
+
+        DeliveryFailure.objects.create(contact_id=wa_id, number_of_failures=5)
+
+        tasks.rapidpro = TembaClient("textit.in", "test-token")
+
+        responses.add(
+            responses.GET,
+            f"https://textit.in/api/v2/contacts.json?uuid={contact_uuid}",
+            json={
+                "results": [
+                    {
+                        "uuid": contact_uuid,
+                        "name": "",
+                        "language": "zul",
+                        "groups": [],
+                        "fields": {},
+                        "blocked": False,
+                        "stopped": False,
+                        "created_on": "2015-11-11T08:30:24.922024+00:00",
+                        "modified_on": "2015-11-11T08:30:25.525936+00:00",
+                        "urns": [f"whatsapp:{wa_id}"],
+                    }
+                ],
+                "next": None,
+            },
+        )
+
+        response = self.client.post(
+            self.url,
+            {
+                "contact_id": contact_uuid,
+                "device_contact_id": "d80d51cb-8a95-4588-ac74-250d739edef8",
+                "id_type": "passport",
+                "id_number": "",
+                "passport_country": "zw",
+                "passport_number": "FN123456",
+                "date_of_birth": None,
+                "language": "zul",
+                "edd": "2020-10-11",
+                "facility_code": "123456",
+                "source": "WhatsApp",
+            },
+            format="json",
+        )
+
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+        [registration] = PrebirthRegistration.objects.all()
+        self.assertEqual(str(registration.contact_id), contact_uuid)
+        self.assertEqual(
+            str(registration.device_contact_id), "d80d51cb-8a95-4588-ac74-250d739edef8"
+        )
+        self.assertEqual(registration.id_type, PASSPORT_IDTYPE)
+        self.assertEqual(registration.id_number, "")
+        self.assertEqual(registration.passport_country, "zw")
+        self.assertEqual(registration.passport_number, "FN123456")
+        self.assertEqual(registration.date_of_birth, None)
+        self.assertEqual(registration.language, "zul")
+        self.assertEqual(registration.edd, datetime.date(2020, 10, 11))
+        self.assertEqual(registration.facility_code, "123456")
+        self.assertEqual(registration.source, "WhatsApp")
+        self.assertEqual(registration.created_by, user.username)
+
+        [delivery_failure] = DeliveryFailure.objects.all()
+        self.assertEqual(delivery_failure.contact_id, wa_id)
+        self.assertEqual(delivery_failure.number_of_failures, 0)
 
     def test_prebirth_other_passport_origin(self):
         """
@@ -1498,7 +1581,8 @@ class Covid19TriageViewSetTests(APITestCase, BaseEventTestCase):
         response = self.client.post(self.url)
         self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
 
-    def test_successful_request(self):
+    @mock.patch("eventstore.views.mark_turn_contact_healthcheck_complete")
+    def test_successful_request(self, task):
         """
         Should create a new Covid19Triage object in the database
         """
@@ -1544,6 +1628,7 @@ class Covid19TriageViewSetTests(APITestCase, BaseEventTestCase):
         self.assertNotEqual(covid19triage.deduplication_id, "")
         self.assertEqual(covid19triage.risk, Covid19Triage.RISK_LOW)
         self.assertEqual(covid19triage.created_by, user.username)
+        task.delay.assert_called_once_with("+27820001001")
 
     def test_duplicate_request(self):
         """
@@ -1784,6 +1869,54 @@ class Covid19TriageV2ViewSetTests(Covid19TriageViewSetTests):
                 "data": {},
             },
         )
+
+    def test_returning_user(self):
+        """
+        Should create a new Covid19Triage object in the database using information
+        from the first entry in the database
+        """
+        user = get_user_model().objects.create_user("test")
+        user.user_permissions.add(Permission.objects.get(codename="add_covid19triage"))
+        Covid19Triage.objects.create(
+            msisdn="+27820001001",
+            province="ZA-WC",
+            city="cape town",
+            fever=False,
+            cough=False,
+            sore_throat=False,
+            tracing=True,
+        )
+        Covid19Triage.objects.create(
+            msisdn="+27820001001",
+            province="ZA-GT",
+            city="sandton",
+            fever=False,
+            cough=False,
+            sore_throat=False,
+            tracing=True,
+        )
+
+        self.client.force_authenticate(user)
+        response = self.client.post(
+            self.url,
+            {
+                "msisdn": "27820001001",
+                "source": "USSD",
+                "age": Covid19Triage.AGE_18T40,
+                "fever": False,
+                "cough": False,
+                "sore_throat": False,
+                "exposure": Covid19Triage.EXPOSURE_NO,
+                "tracing": True,
+                "risk": Covid19Triage.RISK_LOW,
+            },
+            format="json",
+        )
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+        triage_id = response.data["id"]
+        covid19triage = Covid19Triage.objects.get(id=triage_id)
+        self.assertEqual(covid19triage.province, "ZA-WC")
+        self.assertEqual(covid19triage.city, "cape town")
 
 
 class HealthCheckUserProfileViewSetTests(APITestCase, BaseEventTestCase):