+++ eventstore/test_whatsapp_actions.py
@@ -1,11 +1,14 @@
 import datetime
 import json
+from datetime import timedelta
 from unittest.mock import Mock, patch
 
 import responses
 from django.conf import settings
 from django.test import TestCase as DjangoTestCase
 from django.test import override_settings
+from django.utils import timezone
+from django_redis import get_redis_connection
 from pytz import UTC
 from temba_client.v2 import TembaClient
 
@@ -255,9 +258,9 @@ class HandleEventTests(DjangoTestCase):
         event.fallback_channel = True
         event.status = Event.FAILED
         event.recipient_id = "27820001001"
-        event.timestamp = datetime.datetime(2018, 2, 15, 11, 38, 20, tzinfo=UTC)
+        event.timestamp = timezone.now() + timedelta(days=2)
 
-        DeliveryFailure.objects.create(number_of_failures=5, contact_id="27820001001")
+        DeliveryFailure.objects.create(number_of_failures=4, contact_id="27820001001")
 
         with patch("eventstore.tasks.rapidpro") as p:
             handle_fallback_event(event)
@@ -265,7 +268,7 @@ class HandleEventTests(DjangoTestCase):
         p.create_flow_start.assert_called_once_with(
             extra={
                 "optout_reason": "sms_failure",
-                "timestamp": 1518694700,
+                "timestamp": event.timestamp.timestamp(),
                 "babyloss_subscription": "FALSE",
                 "delete_info_for_babyloss": "FALSE",
                 "delete_info_consent": "FALSE",
@@ -274,6 +277,49 @@ class HandleEventTests(DjangoTestCase):
             flow="test-flow-uuid",
             urns=["whatsapp:27820001001"],
         )
+
+    @override_settings(RAPIDPRO_OPTOUT_FLOW="test-flow-uuid")
+    def test_fallback_channel_delivery_failure_duplicate(self):
+        """
+        If the event is of type Failed, and uses the fallback channel, but there
+        was another failure in the last 24h then it should not trigger the
+        message delivery failed action
+        """
+        event = Event.objects.create()
+        event.fallback_channel = True
+        event.status = Event.FAILED
+        event.recipient_id = "27820001001"
+        event.timestamp = timezone.now()
+
+        DeliveryFailure.objects.create(number_of_failures=4, contact_id="27820001001")
+
+        with patch("eventstore.tasks.rapidpro") as p:
+            handle_fallback_event(event)
+
+        p.create_flow_start.assert_not_called()
+        df = DeliveryFailure.objects.get(contact_id="27820001001")
+        self.assertEqual(df.number_of_failures, 4)
+
+    @override_settings(RAPIDPRO_OPTOUT_FLOW="test-flow-uuid")
+    def test_fallback_channel_delivery_failure_error_more_than_5(self):
+        """
+        If the event is of type Failed, and uses the fallback channel,
+        then it should trigger the message delivery failed action
+        """
+        event = Event.objects.create()
+        event.fallback_channel = True
+        event.status = Event.FAILED
+        event.recipient_id = "27820001001"
+        event.timestamp = timezone.now() + timedelta(days=2)
+
+        DeliveryFailure.objects.create(number_of_failures=5, contact_id="27820001001")
+
+        with patch("eventstore.tasks.rapidpro") as p:
+            handle_fallback_event(event)
+
+        p.create_flow_start.assert_not_called()
+        df = DeliveryFailure.objects.get(contact_id="27820001001")
+        self.assertEqual(df.number_of_failures, 6)
 
     def test_fallback_channel_delivery_failure_less_than_5(self):
         """
@@ -285,7 +331,7 @@ class HandleEventTests(DjangoTestCase):
         event.fallback_channel = True
         event.status = Event.FAILED
         event.recipient_id = "27820001001"
-        event.timestamp = datetime.datetime(2018, 2, 15, 11, 38, 20, tzinfo=UTC)
+        event.timestamp = timezone.now() + timedelta(days=2)
 
         with patch("eventstore.tasks.rapidpro") as p:
             handle_fallback_event(event)
@@ -293,6 +339,27 @@ class HandleEventTests(DjangoTestCase):
         p.create_flow_start.assert_not_called()
         df = DeliveryFailure.objects.get(contact_id="27820001001")
         self.assertEqual(df.number_of_failures, 1)
+
+    @override_settings(DISABLE_SMS_FAILURE_OPTOUTS=True)
+    def test_fallback_channel_delivery_failure_optouts_disabled(self):
+        """
+        If the event is of type Failed, and uses the fallback channel,
+        but SMS failure optouts are disabled, it should not call the rapidpro
+        flow
+        """
+        event = Event.objects.create()
+        event.fallback_channel = True
+        event.status = Event.FAILED
+        event.recipient_id = "27820001001"
+        event.timestamp = timezone.now() + timedelta(days=2)
+
+        with patch("eventstore.tasks.rapidpro") as p:
+            handle_fallback_event(event)
+
+        p.create_flow_start.assert_not_called()
+        self.assertFalse(
+            DeliveryFailure.objects.filter(contact_id="27820001001").exists()
+        )
 
     def test_fallback_channel_successful_with_no_existing_delivery_failure(self):
         """
@@ -304,7 +371,7 @@ class HandleEventTests(DjangoTestCase):
         event.fallback_channel = True
         event.status = Event.READ
         event.recipient_id = "27820001001"
-        event.timestamp = datetime.datetime(2018, 2, 15, 11, 38, 20, tzinfo=UTC)
+        event.timestamp = timezone.now() + timedelta(days=2)
 
         with patch("eventstore.tasks.rapidpro") as p:
             handle_fallback_event(event)
@@ -323,7 +390,7 @@ class HandleEventTests(DjangoTestCase):
         event.fallback_channel = True
         event.status = Event.SENT
         event.recipient_id = "27820001001"
-        event.timestamp = datetime.datetime(2018, 2, 15, 11, 38, 20, tzinfo=UTC)
+        event.timestamp = timezone.now() + timedelta(days=2)
 
         DeliveryFailure.objects.create(number_of_failures=3, contact_id="27820001001")
 
@@ -344,7 +411,7 @@ class HandleEventTests(DjangoTestCase):
         event.fallback_channel = True
         event.status = Event.READ
         event.recipient_id = "27820001001"
-        event.timestamp = datetime.datetime(2018, 2, 15, 11, 38, 20, tzinfo=UTC)
+        event.timestamp = timezone.now() + timedelta(days=2)
 
         DeliveryFailure.objects.create(number_of_failures=1, contact_id="27820001001")
 
@@ -375,6 +442,13 @@ class HandleEventTests(DjangoTestCase):
 
 
 class HandleWhatsappEventsTests(DjangoTestCase):
+    def setUp(self):
+        redis = get_redis_connection("redis")
+        key = f"hub_handle_whatsapp_delivery_error_27820001001"
+        redis.delete(key)
+
+        return super().setUp()
+
     @override_settings(RAPIDPRO_UNSENT_EVENT_FLOW="test-flow-uuid")
     @override_settings(ENABLE_UNSENT_EVENT_ACTION=True)
     def test_handle_whatsapp_hsm_error_successful(self):
@@ -393,7 +467,7 @@ class HandleWhatsappEventsTests(DjangoTestCase):
             extra={
                 "popi_ussd": settings.POPI_USSD_CODE,
                 "optout_ussd": settings.OPTOUT_USSD_CODE,
-                "timestamp": 1518694700,
+                "timestamp": 1_518_694_700,
             },
             flow="test-flow-uuid",
             urns=["whatsapp:27820001001"],
@@ -422,7 +496,7 @@ class HandleWhatsappEventsTests(DjangoTestCase):
         Sends a SMS and updates the contact if the contact hasn't been sent this
         message in 30 days
         """
-        timestamp = 1543999390.069308
+        timestamp = 1_543_999_390.069_308
         mock_get_utc_now.return_value = datetime.datetime.fromtimestamp(timestamp)
 
         event = Event.objects.create()
@@ -508,7 +582,7 @@ class HandleWhatsappEventsTests(DjangoTestCase):
         Sends a SMS and updates the contact if the contact hasn't been sent this
         message in 30 days
         """
-        timestamp = 1543999390.069308
+        timestamp = 1_543_999_390.069_308
         mock_get_utc_now.return_value = datetime.datetime.fromtimestamp(timestamp)
 
         event = Event.objects.create()
@@ -595,7 +669,7 @@ class HandleWhatsappEventsTests(DjangoTestCase):
         """
         Doesn't send a SMS if contact recieved the message in the last 30 days
         """
-        timestamp = 1543999390.069308
+        timestamp = 1_543_999_390.069_308
         mock_get_utc_now.return_value = datetime.datetime.fromtimestamp(timestamp)
 
         event = Event.objects.create()
@@ -638,7 +712,7 @@ class HandleWhatsappEventsTests(DjangoTestCase):
         """
         Doesn't fail when the language is None
         """
-        timestamp = 1543999390.069308
+        timestamp = 1_543_999_390.069_308
         mock_get_utc_now.return_value = datetime.datetime.fromtimestamp(timestamp)
 
         event = Event.objects.create()
@@ -681,7 +755,7 @@ class HandleWhatsappEventsTests(DjangoTestCase):
         """
         Doesn't fail when there is no contact
         """
-        timestamp = 1543999390.069308
+        timestamp = 1_543_999_390.069_308
         mock_get_utc_now.return_value = datetime.datetime.fromtimestamp(timestamp)
 
         event = Event.objects.create()